fix(ci): python-full-coverage — split oversized compose files under line limit
Move 8 compose env passthrough/literal tests from
test_workspace_services_compose_part_001 (1583 lines) into a new
test_workspace_services_compose_part_003, and condense the
_compose_environment_mapping docstring in src/awf/profiles/compose.py
(1505 -> 1493 lines). Both files now stay under the 1500-line
first-party limit enforced by test_core_decomposition_maintainability.
